requirement.txt
requests
chardet
bs4
Pillow
pdfkit